pyproject.toml
[tool.poetry]
name = "ai-air-bot"
version = "0.1.0"
description = "Multi-Role ChatGPT Telegram Bot"
authors = ["Ilia Samoletov <airscods@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.0.0b7"
gTTS = "^2.3.2"
loguru = "^0.7.0"
motor = "^3.1.2"
openai = "^0.27.7"
pydub = "^0.25.1"
python-dotenv = "^1.0.0"
streamlit = "^1.22.0"
tiktoken = "^0.4.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
